requirements.txt
telethon
python-dotenv
google-genai
Pillow
